# backendupdate/backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dashboard Backend API",
    "version": "1.0",
    "description": "API для управления дашбордом посещаемости студентов",
    "contact": {
      "name": "API Support",
      "email": "support@dashboard.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Локальный сервер"
    }
  ],
...
  "paths": {
    "/admin/refresh-data": {
      "post": {
        "tags": ["admin"],
        "summary": "Ручное обновление данных",
        "description": "Запускает конвертацию Excel файлов в JSON и загрузку в БД",
        "responses": {
          "200": {
            "description": "Данные успешно обновлены",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "message": { "type": "string", "example": "Данные успешно обновлены" },
                    "time": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Обновление уже выполняется",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Обновление уже выполняется" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Ошибка обновления данных",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" },
                    "details": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}